requirements.txt
aiofiles==0.8.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.4.2
apache-airflow-providers-celery==3.0.0
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-google==8.4.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
appnope==0.1.3
argcomplete==2.0.0
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backoff==2.2.1
billiard==3.6.4.0
blinker==1.5
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
contourpy==1.0.6
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
cycler==0.11.0
db-dtypes==1.0.4
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.2.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flower==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gcloud-aio-auth==4.0.1
gcloud-aio-bigquery==6.1.2
gcloud-aio-storage==7.0.1
google-ads==18.0.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-aiplatform==1.16.1
google-cloud-appengine-logging==1.1.3
google-cloud-audit-log==0.2.4
google-cloud-automl==2.8.0
google-cloud-bigquery==2.34.4
google-cloud-bigquery-datatransfer==3.7.0
google-cloud-bigquery-storage==2.14.1
google-cloud-bigtable==1.7.2
google-cloud-build==3.9.0
google-cloud-container==2.11.1
google-cloud-core==2.3.2
google-cloud-datacatalog==3.9.0
google-cloud-dataform==0.2.0
google-cloud-dataplex==1.1.0
google-cloud-dataproc==5.0.0
google-cloud-dataproc-metastore==1.6.0
google-cloud-dlp==1.0.2
google-cloud-kms==2.12.0
google-cloud-language==1.3.2
google-cloud-logging==3.2.1
google-cloud-memcache==1.4.1
google-cloud-monitoring==2.11.0
google-cloud-orchestration-airflow==1.4.1
google-cloud-os-login==2.7.1
google-cloud-pubsub==2.13.5
google-cloud-redis==2.9.0
google-cloud-resource-manager==1.6.0
google-cloud-secret-manager==1.0.2
google-cloud-spanner==1.19.3
google-cloud-speech==1.3.4
google-cloud-storage==1.44.0
google-cloud-tasks==2.10.1
google-cloud-texttospeech==1.0.3
google-cloud-translate==1.7.2
google-cloud-videointelligence==1.16.3
google-cloud-vision==1.0.2
google-cloud-workflows==1.7.1
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
greenlet==1.1.3.post0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-gcp==0.2.2
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.16.0
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
looker-sdk==22.20.0
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
multidict==6.0.2
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandas-gbq==0.17.9
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.4
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.19.6
protobuf==3.20.0
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.5
pytzdata==2020.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.27
sqlalchemy-bigquery==1.4.4
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
stack-data==0.6.1
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.0.1
text-unidecode==1.3
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
yarl==1.8.1
